[ Refactor ] : Modification based on review
