+++ src/components/Atoms/Inner.ts
@@ -4,8 +4,6 @@ const Inner = styled.div`
     width: 768px;
     margin: 0 auto;
     padding: 0 17px;
-    /* border-left: 2px solid skyblue;
-    border-right: 2px solid skyblue; */
     background-color: #252525;
 `;
