+++ src/components/MemberSet/MemberForm.ts
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 
 const MemberForm = styled.form`
-    padding-top: 7vh;
+    padding-top: 50px;
 `;
 
 export default MemberForm;
