+++ src/components/Profile/index.ts
@@ -1,1 +1,1 @@
-export { default as Profile } from './Profile';
+export { default } from './Profile';
