--- src/components/common/index.ts
@@ -1,2 +1,0 @@
-export { default as Inner } from './Inner';
-export { default as Main } from './Main';
